Move run lib to components and update imports
Refactor internal.run package by renaming and moving files from run/lib and run/config into run/components (including asynctask, builder, consumer, null, config/* and fmt). Update imports across the package and tests to reference swanlab.sdk.internal.run.components and swanlab.sdk.internal.run.fmt; remove the old run/lib __init__.py. Also expose the run components' config at the top-level SDK init and adjust related tests/benchmarks and conftest to match the new module paths.
