+++ tests/unit/conftest.py
@@ -5,7 +5,7 @@ from swanlab.sdk.internal.context import callbacker
 from swanlab.sdk.internal.core_python import client
 from swanlab.sdk.internal.pkg import console
 from swanlab.sdk.internal.run import clear_run
-from swanlab.sdk.internal.run.config import reset as reset_config
+from swanlab.sdk.internal.run.components.config import reset as reset_config
 from swanlab.sdk.internal.settings import Settings, settings
