+++ tests/unit/sdk/internal/run/test_run_fmt.py
@@ -7,7 +7,7 @@
 
 import pytest
 
-import swanlab.sdk.internal.run.lib.fmt as fmt
+import swanlab.sdk.internal.run.fmt as fmt
 
 # ─────────────────────────── flatten_dict ────────────────────────────
 
@@ -43,7 +43,7 @@ class TestFlattenDict:
     def test_duplicate_key_latter_wins(self, monkeypatch):
         """出现同名展开键时，后者覆盖前者，并触发一次警告"""
         mock_warning = []
-        monkeypatch.setattr("swanlab.sdk.internal.run.lib.fmt.console.warning", lambda *args: mock_warning.append(args))
+        monkeypatch.setattr("swanlab.sdk.internal.run.fmt.console.warning", lambda *args: mock_warning.append(args))
 
         data = {"a": {"b": 1}, "a/b": 99}
         result = fmt.flatten_dict(data)
@@ -79,7 +79,7 @@ class TestValidateKey:
 
     def test_strip_leading_trailing(self, monkeypatch):
         """头尾的空白、. 和 / 被剥离"""
-        monkeypatch.setattr("swanlab.sdk.internal.run.lib.fmt.console.warning", lambda *args: None)
+        monkeypatch.setattr("swanlab.sdk.internal.run.fmt.console.warning", lambda *args: None)
         assert fmt.validate_key("  loss  ") == "loss"
         assert fmt.validate_key("./key/") == "key"
         assert fmt.validate_key("...key...") == "key"
@@ -110,7 +110,7 @@ class TestValidateKey:
 
     def test_custom_max_len(self, monkeypatch):
         """自定义 max_len 生效"""
-        monkeypatch.setattr("swanlab.sdk.internal.run.lib.fmt.console.warning", lambda *args: None)
+        monkeypatch.setattr("swanlab.sdk.internal.run.fmt.console.warning", lambda *args: None)
         assert fmt.validate_key("abcdef", max_len=3) == "abc"
 
     @pytest.mark.parametrize("key, expected", [(42, "42"), (3.14, "3.14")])
@@ -128,7 +128,7 @@ class TestValidateKey:
     def test_warning_fires_once_per_key(self, monkeypatch):
         """相同 key 被头尾剥离时，只触发一次警告"""
         mock_warning = []
-        monkeypatch.setattr("swanlab.sdk.internal.run.lib.fmt.console.warning", lambda *args: mock_warning.append(args))
+        monkeypatch.setattr("swanlab.sdk.internal.run.fmt.console.warning", lambda *args: mock_warning.append(args))
 
         fmt.validate_key("  bad_key  ")
         fmt.validate_key("  bad_key  ")
@@ -139,7 +139,7 @@ class TestValidateKey:
     def test_warning_fires_for_different_keys(self, monkeypatch):
         """不同被剥离的 key 各自独立触发一次警告"""
         mock_warning = []
-        monkeypatch.setattr("swanlab.sdk.internal.run.lib.fmt.console.warning", lambda *args: mock_warning.append(args))
+        monkeypatch.setattr("swanlab.sdk.internal.run.fmt.console.warning", lambda *args: mock_warning.append(args))
 
         fmt.validate_key("  bad_key1  ")
         fmt.validate_key("  bad_key2  ")
